user-search-backend/pyproject.toml
[tool.poetry]
name = "user-search-backend"
version = "0.1.0"
description = "User profile search backend with Flask"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
flask = "^2.3.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.10"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
ipykernel = "^6.30.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
